use new make data files

+++ src/utils/dataset_stats.py
@@ -1,52 +1,72 @@
-"""Functions used for generating the band statistics and the fmask label
+"""Functions used for generating the label
 statistics csv files to be used during the training. """
 
-import csv
+import os, sys
 import glob
-import os
-
-import h5py
 import numpy as np
+import csv
+import h5py
+import argparse
+import logging
+from sklearn.metrics import confusion_matrix
+logger = logging.getLogger('__name__')
 
 
-def _write_to_csv(csv_path, fields, values_dict):
-    """Writes the given values_dict to csv file.
-    If the file does not already exist, a header entry is also added using the
-    fields argument."""
+def setup_logger():
+    logging.basicConfig(
+        level=logging.INFO,
+        format='%(asctime)s %(message)s',
+    )
 
-    if os.path.exists(csv_path):
-        mode = 'a'
+
+def get_args(argv=None):
+    """Parses the arguments entered by the user."""
+
+    parser = argparse.ArgumentParser(description="Gives stats of H5 files ")
+    parser.add_argument('-p', '--path', help='Path of H5 files',
+                        default=os.getcwd())
+    parser.add_argument('-m', '--mode', default='train',
+                        help='mode for data use')
+
+    return parser.parse_args(argv)
+
+
+
+
+np.set_printoptions(precision=4, suppress=True)
+
+
+def count_label(label_mask):
+    uniques, counts = np.unique(label_mask, return_counts=True)
+    label_dict = {i: 0 for i in range(6)}
+    for unique, countr in zip(uniques, counts):
+        label_dict[unique] = countr
+    return label_dict
+
+
+def get_label_info(data, mode):
+    fmask = data[:, :, 13]
+
+    fmask_count = count_label(fmask)
+    if mode == 'train':
+        true_count = None
+        conf_mat = np.zeros((6, 6))
     else:
-        mode = 'w'
-    with open(csv_path, mode=mode) as csv_file:
-        writer = csv.DictWriter(csv_file, fieldnames=fields)
-        if mode == 'w':
-            writer.writeheader()
-        writer.writerow(values_dict)
+        true_label = data[:, :, 14]
+
+        true_count = count_label(true_label)
+        true_count[0] = 0
+
+        true_label_filtered = true_label[true_label != 0].reshape(-1, )
+        fmask_filtered = fmask[true_label != 0].reshape(-1, )
+
+        conf_mat = confusion_matrix(true_label_filtered, fmask_filtered,
+                                    labels=[0, 1, 2, 3, 4, 5])
+
+    return fmask_count, true_count, conf_mat
 
 
-def _make_label_entry(filename, fmask_count, label_csv_path):
-    """Generates csv entry for each row of the label stats file"""
-    fields = ['FILENAME', 'NONE_F', 'CLEAR_F',
-              'CLOUD_F', 'SHADOW_F', 'ICE_F',
-              'WATER_F']
-    values = [filename]
-    values.extend(fmask_count)
-    values_dict = dict(zip(fields, values))
-    _write_to_csv(label_csv_path, fields, values_dict)
-
-
-def _count_label(label_mask):
-    """Counts the number pixels for each class label"""
-    labels, counts = np.unique(label_mask, return_counts=True)
-    count_array=np.zeros(6,dtype=np.uint64)
-    for label, countr in zip(labels, counts):
-        count_array[label] = countr
-    return count_array
-
-
-def _get_band_info(data):
-    """Computes statistics for each band"""
+def get_band_info(data):
     bands = data[:, :, :13]
     band_mean = np.mean(bands, axis=(0, 1))
     band_std = np.std(bands, axis=(0, 1))
@@ -56,8 +76,32 @@ def _get_band_info(data):
     return [band_mean, band_std, band_min, band_max]
 
 
-def _make_band_entry(filename, stats, band_csv_path):
-    """Generates csv entry for each row of the band stats file"""
+def make_label_entry(filename, fmask_count, val_count, label_csv_path):
+    """Generates csv entry for each row of the label stats file"""
+    fields = ['FILENAME', 'NONE_F', 'CLEAR_F',
+              'CLOUD_F', 'SHADOW_F', 'ICE_F',
+              'WATER_F']
+    values = [filename]
+    values.extend(fmask_count.values())
+    if val_count is not None:
+        fields.extend(['NONE_T', 'CLEAR_T',
+                       'CLOUD_T', 'SHADOW_T', 'ICE_T',
+                       'WATER_T'])
+        values.extend(val_count.values())
+    values_dict = dict(zip(fields, values))
+
+    if os.path.exists(label_csv_path):
+        mode = 'a'
+    else:
+        mode = 'w'
+    with open(label_csv_path, mode=mode) as csv_file:
+        writer = csv.DictWriter(csv_file, fieldnames=fields)
+        if mode == 'w':
+            writer.writeheader()
+        writer.writerow(values_dict)
+
+
+def make_band_entry(filename, stats, band_csv_path):
     N_BANDS = 13
     fields = ['FILENAME', ]
     for stat_name in ['mean', 'std', 'min', 'max']:
@@ -67,60 +111,74 @@ def _make_band_entry(filename, stats, band_csv_path):
         values.extend(stats[i].tolist())
 
     values_dict = dict(zip(fields, values))
-    _write_to_csv(band_csv_path, fields, values_dict)
+
+    if os.path.exists(band_csv_path):
+        mode = 'a'
+    else:
+        mode = 'w'
+    with open(band_csv_path, mode=mode) as csv_file:
+        writer = csv.DictWriter(csv_file, fieldnames=fields)
+        if mode == 'w':
+            writer.writeheader()
+        writer.writerow(values_dict)
 
 
-def save_stats(h5_path):
-    """Main function that generates the statistics csv files for all the h5 files
-    in the given path """
-    current_path = os.path.abspath(h5_path)
+def main_csv(mode, path):
+    current_path = os.path.abspath(path)
+    logger.info('Running dataset stats for {} in {} mode'.format(current_path, mode))
 
     H5_files = glob.glob(os.path.join(current_path, "*.h5"))
     label_csv_path = os.path.join(current_path, 'label_stats.csv')
     band_csv_path = os.path.join(current_path, 'band_stats.csv')
-
     if os.path.exists(label_csv_path):
         os.remove(label_csv_path)
-    if os.path.exists(band_csv_path):
+    if mode == 'train' and os.path.exists(band_csv_path):
         os.remove(band_csv_path)
-
-    total_fmask_count = np.zeros(6, dtype=np.uint64)
-    full_dataset_stats = {
-        'mean': np.zeros(13),
-        'std': np.zeros(13),
-        'min': 500000 * np.ones(13),
-        'max': np.zeros(13),
-        'count': 0
-    }
+    stats = np.zeros((6))
+    conf_mat = np.zeros((6, 6))
 
     for h5_file_path in H5_files:
         with h5py.File(h5_file_path, 'r') as hf:
             spectral_image = hf.get('data')[:]
-        filename = os.path.basename(h5_file_path)
-        fmask_count = _count_label(spectral_image[:, :, 13])
+            filename = os.path.basename(h5_file_path)
+            fmask_count, true_count, conf_ = get_label_info(spectral_image,
+                                                            mode)
 
-        band_stats = _get_band_info(spectral_image)
+            if mode == 'test':
+                conf_mat += conf_
+                for i in true_count:
+                    stats[i] += true_count[i]
+            else:
+                for i in fmask_count:
+                    stats[i] += fmask_count[i]
+            make_label_entry(filename, fmask_count, true_count, label_csv_path)
 
-        _make_band_entry(filename, band_stats, band_csv_path)
-        _make_label_entry(filename, fmask_count, label_csv_path)
+    if mode == 'test':
 
-        total_fmask_count += fmask_count
-        full_dataset_stats['mean'] += band_stats[0]
-        full_dataset_stats['std'] += np.square(band_stats[1])
-        full_dataset_stats['min'] = np.minimum(band_stats[2],
-                                            full_dataset_stats['min'])
-        full_dataset_stats['max'] = np.maximum(band_stats[3],
-                                            full_dataset_stats['max'])
-        full_dataset_stats['count'] += 1
+        if np.sum(conf_mat[1:, 1:]):
+            logger.info('Confusion matrix comparing FMask and true labels \n{}'.format(conf_mat.astype(int)))
+            fmask_correct = np.sum(conf_mat.diagonal()) / np.sum(conf_mat)
+            logger.info('Fmask Agreement with labels=> {:.3%}'.format(fmask_correct))
+        else:
+            logger.info('No FMask labels found')
+
+        out_string = 'NONE:{:.2%} \t CLEAR:{:.2%} \t CLOUD:{:.2%} \t' \
+                     'SHADOW:{:.2%} \t ICE:{:.2%} \t ' \
+                     'WATER:{:.2%}'.format(*stats / np.sum(stats))
+        logger.info('Test Class Distribution => {}'.format(out_string))
 
 
-    # Aggregate complete dataset statistics
-    full_band_list = [full_dataset_stats['mean'] / full_dataset_stats['count'],
-                      np.sqrt(full_dataset_stats['std'] / full_dataset_stats['count']),
-                      full_dataset_stats['min'], full_dataset_stats['max']]
+    else:
+        out_string = 'NONE:{:.2%} \t CLEAR:{:.2%} \t CLOUD:{:.2%} \t' \
+                     'SHADOW:{:.2%} \t ICE:{:.2%} \t ' \
+                     'WATER:{:.2%}'.format(*stats / np.sum(stats))
+        logger.info('Train Class Distribution => {}'.format(out_string))
 
-    _make_band_entry('COMPLETE', full_band_list, band_csv_path)
+
+
 
 
 if __name__ == '__main__':
-    pass
+    setup_logger()
+    args = get_args()
+    main_csv(mode=args.mode, path=args.path)
